rename package `de.vzg.maven` to `org.mycore.plugins.maven.solr` fixes #3
