+++ pom.xml
@@ -24,7 +24,7 @@
   <parent>
     <groupId>org.mycore</groupId>
     <artifactId>mycore-parent</artifactId>
-    <version>37-SNAPSHOT</version>
+    <version>38</version>
   </parent>
   <groupId>org.mycore.plugins</groupId>
   <artifactId>solr-runner-maven-plugin</artifactId>
