--- src/main/java/de/vzg/maven/PluginCoreMapping.java
@@ -1,15 +1,0 @@
-package de.vzg.maven;
-
-public class PluginCoreMapping {
-
-    private String core;
-    private String plugin;
-
-    public String getCore() {
-        return core;
-    }
-
-    public String getPlugin() {
-        return plugin;
-    }
-}
